refactor(NcIconSvgWrapper): migrate component to Typescript and script-setup
