+++ src/components/NcIconSvgWrapper/NcIconSvgWrapper.vue
@@ -158,6 +158,80 @@ export default {
 ```
 </docs>
 
+<script setup lang="ts">
+import { computed, warn } from 'vue'
+import DOMPurify from 'dompurify'
+
+const props = withDefaults(defineProps<{
+	/**
+	 * Make the icon directional, meaning it is langauge direction aware.
+	 * If the icon is placed in a right-to-left context it will be mirrored vertically.
+	 */
+	directional?: boolean
+
+	/**
+	 * Set if the icon should be used as inline content e.g. within text.
+	 * By default the icon is made a block element for use inside `icon`-slots.
+	 */
+	inline?: boolean
+
+	/**
+	 * Raw SVG string to render
+	 */
+	svg?: string
+
+	/**
+	 * Label of the icon, used in aria-label
+	 */
+	name?: string
+
+	/**
+	 * Raw SVG path to render. Takes precedence over the SVG string in the `svg` prop.
+	 */
+	path?: string
+
+	/**
+	 * Size of the icon to show. Only use if not using within an icon slot.
+	 * Defaults to 20px which is the Nextcloud icon size for all icon slots.
+	 */
+	size?: number | 'auto'
+}>(), {
+	name: undefined,
+	path: '',
+	size: 20,
+	svg: '',
+})
+
+/**
+ * Icon size used in CSS
+ */
+const iconSize = computed(() => typeof props.size === 'number' ? `${props.size}px` : props.size)
+
+/**
+ * The sanitized SVG or undefined if path shall be used
+ */
+const cleanSvg = computed(() => {
+	if (!props.svg || props.path) {
+		return
+	}
+
+	const svg = DOMPurify.sanitize(props.svg)
+
+	const svgDocument = new DOMParser().parseFromString(svg, 'image/svg+xml')
+
+	if (svgDocument.querySelector('parsererror')) {
+		warn('SVG is not valid')
+		return ''
+	}
+
+	if (svgDocument.documentElement.id) {
+		svgDocument.documentElement.removeAttribute('id')
+	}
+
+	return svgDocument.documentElement.outerHTML
+})
+</script>
+
 <template>
 	<span :aria-hidden="name ? undefined : 'true'"
 		:aria-label="name || undefined"
@@ -174,100 +248,6 @@ export default {
 		<span v-else v-html="cleanSvg" />
 	</span>
 </template>
-
-<script>
-import { warn } from 'vue'
-import DOMPurify from 'dompurify'
-
-export default {
-	name: 'NcIconSvgWrapper',
-
-	props: {
-		/**
-		 * Make the icon directional, meaning it is langauge direction aware.
-		 * If the icon is placed in a right-to-left context it will be mirrored vertically.
-		 */
-		directional: {
-			type: Boolean,
-			default: false,
-		},
-
-		/**
-		 * Set if the icon should be used as inline content e.g. within text.
-		 * By default the icon is made a block element for use inside `icon`-slots.
-		 */
-		inline: {
-			type: Boolean,
-			default: false,
-		},
-
-		/**
-		 * Raw SVG string to render
-		 */
-		svg: {
-			type: String,
-			default: '',
-		},
-
-		/**
-		 * Label of the icon, used in aria-label
-		 */
-		name: {
-			type: String,
-			default: '',
-		},
-
-		/**
-		 * Raw SVG path to render. Takes precedence over the SVG string in the `svg` prop.
-		 */
-		path: {
-			type: String,
-			default: '',
-		},
-
-		/**
-		 * Size of the icon to show. Only use if not using within an icon slot.
-		 * Defaults to 20px which is the Nextcloud icon size for all icon slots.
-		 * @default 20
-		 */
-		size: {
-			type: [Number, String],
-			default: 20,
-			validator: (value) => typeof value === 'number' || value === 'auto',
-		},
-	},
-
-	computed: {
-		/**
-		 * Icon size used in CSS
-		 */
-		iconSize() {
-			return typeof this.size === 'number' ? `${this.size}px` : this.size
-		},
-
-		cleanSvg() {
-			if (!this.svg || this.path) {
-				return
-			}
-
-			const svg = DOMPurify.sanitize(this.svg)
-
-			const svgDocument = new DOMParser().parseFromString(svg, 'image/svg+xml')
-
-			if (svgDocument.querySelector('parsererror')) {
-				warn('SVG is not valid')
-				return ''
-			}
-
-			if (svgDocument.documentElement.id) {
-				svgDocument.documentElement.removeAttribute('id')
-			}
-
-			return svgDocument.documentElement.outerHTML
-		},
-	},
-}
-</script>
 
 <style lang="scss" scoped>
 .icon-vue {
