+++ src/components/NcIconSvgWrapper/index.ts
