+++ src/components/NcListItemIcon/NcListItemIcon.vue
@@ -145,7 +145,7 @@ It might be used for list rendering or within the multiselect for example
 <script>
 import NcAvatar from '../NcAvatar/index.js'
 import NcHighlight from '../NcHighlight/index.js'
-import NcIconSvgWrapper from '../NcIconSvgWrapper/index.js'
+import NcIconSvgWrapper from '../NcIconSvgWrapper/index.ts'
 
 import { userStatus } from '../../mixins/index.js'
