+++ src/components/index.ts
@@ -58,7 +58,7 @@ export { default as NcGuestContent } from './NcGuestContent/index.ts'
 export { default as NcHeaderButton } from './NcHeaderButton/index.ts'
 export { default as NcHeaderMenu } from './NcHeaderMenu/index.js'
 export { default as NcHighlight } from './NcHighlight/index.js'
-export { default as NcIconSvgWrapper } from './NcIconSvgWrapper/index.js'
+export { default as NcIconSvgWrapper } from './NcIconSvgWrapper/index.ts'
 export { default as NcInputField } from './NcInputField/index.js'
 export { default as NcListItem } from './NcListItem/index.js'
 export { default as NcListItemIcon } from './NcListItemIcon/index.js'
